# api/swagger/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionInfo": {
        "properties": {
          "cid": {
            "type": "integer"
          },
          "connected_at": {
            "type": "string"
          },
          "in_bytes": {
            "type": "integer"
          },
          "in_msgs": {
            "type": "integer"
          },
          "ip": {
            "type": "string"
          },
          "last_activity": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "out_bytes": {
            "type": "integer"
          },
          "out_msgs": {
            "type": "integer"
          },
          "pending_bytes": {
            "type": "integer"
          },
          "port": {
            "type": "integer"
          },
          "rtt": {
            "type": "string"
          },
          "server": {
            "type": "string"
          },
          "server_id": {
            "type": "string"
          },
          "subs_count": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ConnectionsResponse": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionInfo"
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DashboardStatsResponse": {
        "properties": {
          "bytes": {
            "type": "integer"
          },
          "connections": {
            "type": "integer"
          },
          "consumers": {
            "type": "integer"
          },
          "messages": {
            "type": "integer"
          },
          "server_status": {
            "type": "string"
          },
          "streams": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PaginatedMessagesResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/StreamMessage"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "stream": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "StreamMessage": {
        "properties": {
          "data": {
            "type": "string"
          },
          "data_base64": {
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "sequence": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "subject": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "StreamMessagesResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/StreamMessage"
            },
            "type": "array"
          },
          "stream": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SubjectInfo": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "last_seen": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SubjectsResponse": {
        "properties": {
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectInfo"
            },
            "type": "array"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "NATS Horizon API",
    "version": "1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/account/info": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Account information"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get JetStream account information",
        "tags": [
          "health"
        ]
      }
    },
...
    "/connections": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get NATS connections",
        "tags": [
          "health"
        ]
      }
    },
    "/connections/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Connection ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Terminate a connection",
        "tags": [
          "health"
        ]
      }
    },
...
    "/dashboard/stats": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsResponse"
                }
              }
            },
            "description": "OK"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get dashboard statistics",
        "tags": [
          "health"
        ]
      }
    },
...
    "/messages": {
      "get": {
        "description": "Lists messages from a stream",
        "parameters": [
          {
            "description": "Stream name",
            "in": "query",
            "name": "stream",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of messages to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 25,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamMessagesResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get stream messages",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/page": {
      "get": {
        "description": "Returns paginated messages from a stream.",
        "parameters": [
          {
            "description": "Stream name",
            "in": "query",
            "name": "stream",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "page_size",
            "schema": {
              "default": 25,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get paginated stream messages",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/server/info": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Server information"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get server information",
        "tags": [
          "health"
        ]
      }
    },
...
    "/subjects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectsResponse"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get subject information",
        "tags": [
          "health"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}